Refactor container folder names from teams to organizations
